Move mutateInternal into an internal method
